ADGS1408: Initial commit
The ADGS1408 contains four independent single pole/single-throw switches,
with a SPI interface for communication.
This driver also supports the adgs1409 and daisy chain mode.

It is derived from adgs5412, with Round Robin functionality added for
enabling individual switches just with a toggle on the CNV pin and no other
serial communication.
It also follows the linux style formatting.

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>